+++ src/components/game/index.ts
@@ -1,4 +1,5 @@
 export { default as GameCard } from './GameCard.vue'
 export { default as CardGrid } from './CardGrid.vue'
 export { default as CardContextMenu } from './CardContextMenu.vue'
+export { default as CardSearch } from './CardSearch.vue'
 export { default as SpyCard } from './SpyCard.vue'